+++ engine/src/commonMain/kotlin/org/pcsoft/framework/simplay/engine/SimpLayEngine.kt
@@ -10,9 +10,9 @@
  * See the License for the specific language governing permissions and limitations.
  */
 
-package org.pcsoft.framework.simplay.engine.engine
+package org.pcsoft.framework.simplay.engine
 
-import org.pcsoft.framework.simplay.engine.engine.internal.SimpLayPageEngine
+import org.pcsoft.framework.simplay.engine.internal.SimpLayPageEngine
 import org.pcsoft.framework.simplay.engine.measure.MeasuredDocument
 import org.pcsoft.framework.simplay.engine.model.Document