+++ engine/src/commonMain/kotlin/org/pcsoft/framework/simplay/engine/model/FontStyle.kt
@@ -13,7 +13,6 @@
 package org.pcsoft.framework.simplay.engine.model
 
 
-
 /**
  * Slant of a font face.
  */