+++ engine/src/commonTest/kotlin/org/pcsoft/framework/simplay/engine/MeasuredSizingTest.kt
@@ -10,7 +10,7 @@
  * See the License for the specific language governing permissions and limitations.
  */
 
-package org.pcsoft.framework.simplay.engine.engine
+package org.pcsoft.framework.simplay.engine
 
 import kotlin.test.Test
 import kotlin.test.assertEquals