+++ fx/src/main/kotlin/org/pcsoft/framework/simplay/fx/CanvasRenderConfiguration.kt
@@ -12,7 +12,7 @@
 
 package org.pcsoft.framework.simplay.fx
 
-import org.pcsoft.framework.simplay.engine.engine.RenderConfiguration
+import org.pcsoft.framework.simplay.engine.RenderConfiguration
 
 /**
  * Mutable configuration for [CanvasDocumentRenderer]. Extends the shared, renderer-agnostic