+++ fx/src/test/kotlin/org/pcsoft/framework/simplay/fx/TextSelectionTest.kt
@@ -14,7 +14,7 @@ package org.pcsoft.framework.simplay.fx
 
 import javafx.scene.input.Clipboard
 import org.junit.jupiter.api.Test
-import org.pcsoft.framework.simplay.engine.engine.measure
+import org.pcsoft.framework.simplay.engine.measure
 import org.pcsoft.framework.simplay.engine.geometry.Margins
 import org.pcsoft.framework.simplay.engine.geometry.Size
 import org.pcsoft.framework.simplay.engine.model.*